+++ app/login/qrlogin.ts
@@ -1,10 +1,14 @@
-import { LoginService } from '@/wasm/pkg';
+import { CheckQRCodeLoginResponse, LoginService } from '@/wasm/pkg';
 
 const getQRLoginTicket = async () => {
 	return await LoginService.getQRCodeLoginTicket();
 };
 
-const checkQRLoginState = async ({ ticket }: { ticket: string }) => {
+const checkQRLoginState = async ({
+	ticket,
+}: {
+	ticket: string;
+}): Promise<CheckQRCodeLoginResponse> => {
 	return await LoginService.checkQRCodeLogin(ticket);
 };
