+++ wasm/pkg/package.json
@@ -1,6 +1,5 @@
 {
   "name": "ham-web-wasm",
-  "type": "module",
   "collaborators": [
     "orangeboyChen <chenenhan@qq.com>"
   ],
@@ -11,7 +10,7 @@
     "ham_web_wasm_bg.js",
     "ham_web_wasm.d.ts"
   ],
-  "main": "ham_web_wasm.js",
+  "module": "ham_web_wasm.js",
   "types": "ham_web_wasm.d.ts",
   "sideEffects": [
     "./ham_web_wasm.js",
